added java files
ADP_calc.java is a Java program that connects to an Oracle database and retrieves data from two tables: ROS_2023 and ADP_2024. It calculates the difference between the ROS_ADP (Rest of Season Average Draft Position) from the ROS_2023 table and the rank (RK) from the ADP_2024 table for each player. If ROS_ADP is null, it displays "N/A" and sets the difference to -1. Finally, it prints the player's rank, name, ROS_ADP (or "N/A"), and the difference.

OracleDatabaseConnection.java is a quick test to connect Oracle database with ide and test tables